+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project pretends to adhere to [Semantic Versioning](http://semver.org/)
 
 ### Added
 - Added a `--no-timer` option to silence the timer ([mikaylathompson](https://github.com/mikaylathompson) -- [#9](https://github.com/liviu-/ding/pull/9))
+- Added a `--command` option to optionally replace the `sys.stdout.write('\a')`. This way, one may use a custom command to be called when the timer finishes ([#8](https://github.com/liviu-/ding/issues/8), [#6](https://github.com/liviu-/ding/pull/6), [#5](https://github.com/liviu-/ding/issues/5))
 
 ## [1.3.0] - 2016-10-04
 
@@ -43,6 +44,7 @@ and this project pretends to adhere to [Semantic Versioning](http://semver.org/)
 
 Initial stable release
 
+[Unreleased]: https://github.com/liviu-/ding/compare/v1.3.0..HEAD
 [1.3.0]: https://github.com/liviu-/ding/compare/v1.2.0..v1.3.0
 [1.2.0]: https://github.com/liviu-/ding/compare/v1.1.0..v1.2.0
 [1.1.0]: https://github.com/liviu-/ding/compare/v1.0.0...v1.1.0